chore: rename AutoABR.ts -> AutoVMAF.ts
